requirements_test.txt
flake8
pytest
mysql-connector-python